feat: 🎸 Header cell click event moved, use plain HTML for icon
The event binder to sort the column was moved from the icon's `span` to
the `th` itself. The span now contains plain HTML to display the sort
state, instead of the classes previously used. The header alignment can
now independently be controlled, using the property `headerAlign` in the
`Column` definition.

Issues: #35, #26