refactor: 💡 Add support of async handlers for rows processing
Renamed `processedRows` to `displayedRows`, `filterBy` to `filter`.
Removed the computed `rows`, because it is used in a single place, and
would fail if non array/string/enumerable is passed as data. Handler
methods or `data` calls are `await`ed, so that they can be async. The
`displayHandler` & `data` function calls expect to return an object
containing the `rows` and the `totalRowCount` properties, that may be
Nil to avoid being set on the component.

BREAKING CHANGE: Changed signature of display handler, changed `filterBy` to `filter`,
changed `processedRows` to `displayedRows`.